+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Add: isInitDatetime() method for judging whether specified attribute is init date & time. (vistart)
 - Chg: isInitDatetime() considers the null value is init value. (vistart)
 - Add: $initDatetime & $initTimestamp static properties. (vistart);
+- Add: offsetDatetime() method for generating current or specified date & time with specified offset. (vistart)
+- Add: getUpdatedAt() method, known as magic-property. (vistart)
 
 #### traits/PasswordTrait.php